Replace internal model with API representation model
