+++ org.emoflon.ibex.gt/model/Gt.ecore
@@ -3,99 +3,37 @@
     xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" name="GTLanguage" nsURI="platform:/resource/org.emoflon.ibex.gt/model/Gt.ecore"
     nsPrefix="GTLanguage">
   <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-    <details key="documentation" value="Internal model for graph transformation rules."/>
+    <details key="documentation" value="Model for the Graph Transformation API."/>
   </eAnnotations>
-  <eClassifiers xsi:type="ecore:EClass" name="GTAttribute" abstract="true">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="Super class for Attribute assignments and conditions."/>
-    </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="type" eType="ecore:EClass http://www.eclipse.org/emf/2002/Ecore#//EAttribute"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="value" eType="ecore:EClass ../../org.emoflon.ibex.common/model/Common.ecore#//IBeXAttributeValue"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTAttributeAssignment" eSuperTypes="#//GTAttribute">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="An attribute assignment sets an attribute value."/>
-    </eAnnotations>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTAttributeCondition" eSuperTypes="#//GTAttribute">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="An attribute condition asserts attribute values based on a comparison of the current attribute value with another value."/>
-    </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="relation" eType="ecore:EEnum ../../org.emoflon.ibex.common/model/Common.ecore#//IBeXRelation"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EEnum" name="GTBindingType">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="The binding type defines whether a graph element is context, created or deleted."/>
-    </eAnnotations>
-    <eLiterals name="CONTEXT"/>
-    <eLiterals name="CREATE" value="1" literal="CREATE"/>
-    <eLiterals name="DELETE" value="2" literal="DELETE"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTEdge" eSuperTypes="#//GTGraphElement">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="A typed edge connects two nodes."/>
-    </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="type" lowerBound="1" eType="ecore:EClass http://www.eclipse.org/emf/2002/Ecore#//EReference"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="targetNode" eType="#//GTNode"
-        eOpposite="#//GTNode/incomingEdges"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="sourceNode" eType="#//GTNode"
-        eOpposite="#//GTNode/outgoingEdges"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTGraph">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="A typed graph contains typed nodes and typed edges between the nodes."/>
-    </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="nodes" upperBound="-1"
-        eType="#//GTNode" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="edges" upperBound="-1"
-        eType="#//GTEdge" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTGraphElement" eSuperTypes="#//GTNamedElement">
-    <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="Any element in the internal GT model which is contained in a graph should inherit from this class. A graph element must be of one of the binding types CONTEXT, CREATE or DELETE."/>
-    </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="bindingType" lowerBound="1"
-        eType="#//GTBindingType"/>
-  </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="GTNamedElement">
     <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
       <details key="documentation" value="Any element in the internal model which has a name should inherit form this class."/>
     </eAnnotations>
     <eStructuralFeatures xsi:type="ecore:EAttribute" name="name" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EString"/>
   </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GTNode" eSuperTypes="#//GTGraphElement">
+  <eClassifiers xsi:type="ecore:EClass" name="GTNode" eSuperTypes="#//GTNamedElement">
     <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
       <details key="documentation" value="A typed node can reference other nodes via edges."/>
     </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="attributeAssignments" upperBound="-1"
-        eType="#//GTAttributeAssignment" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="attributeConditions" upperBound="-1"
-        eType="#//GTAttributeCondition" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="incomingEdges" upperBound="-1"
-        eType="#//GTEdge" eOpposite="#//GTEdge/targetNode"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="outgoingEdges" upperBound="-1"
-        eType="#//GTEdge" eOpposite="#//GTEdge/sourceNode"/>
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="local" lowerBound="1" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EBoolean"
-        defaultValueLiteral="false"/>
     <eStructuralFeatures xsi:type="ecore:EReference" name="type" eType="ecore:EClass http://www.eclipse.org/emf/2002/Ecore#//EClass"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="GTParameter" eSuperTypes="#//GTNamedElement">
     <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="Parameters are required arguments for a rule."/>
+      <details key="documentation" value="Parameters are required typed arguments for a rule."/>
     </eAnnotations>
     <eStructuralFeatures xsi:type="ecore:EReference" name="type" eType="ecore:EClass http://www.eclipse.org/emf/2002/Ecore#//EDataType"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="GTRule" eSuperTypes="#//GTNamedElement">
     <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
-      <details key="documentation" value="Rules consists of a graph with typed nodes and edges."/>
+      <details key="documentation" value="A rule contains parameters (optional) and nodes. Only context and deleted nodes can be bound on a rule."/>
     </eAnnotations>
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="abstract" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EBoolean"/>
     <eStructuralFeatures xsi:type="ecore:EAttribute" name="executable" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EBoolean"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="graph" eType="#//GTGraph"
-        containment="true"/>
+    <eStructuralFeatures xsi:type="ecore:EReference" name="nodes" upperBound="-1"
+        eType="#//GTNode" containment="true"/>
     <eStructuralFeatures xsi:type="ecore:EReference" name="parameters" upperBound="-1"
         eType="#//GTParameter" containment="true"/>
+    <eStructuralFeatures xsi:type="ecore:EReference" name="ruleNodes" upperBound="-1"
+        eType="#//GTNode"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="GTRuleSet">
     <eAnnotations source="http://www.eclipse.org/emf/2002/GenModel">
